Introduce and/or compute functions
